socialite_admin/requirements.txt
anyio==3.6.2
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
dj-database-url==1.3.0
Django==4.2
django-environ==0.10.0
djangorestframework==3.14.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
psycopg2-binary==2.9.6
pytz==2023.3
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
urllib3==2.0.2
djangorestframework-simplejwt
django-cors-headers
django-extensions
django-filter
